Allow http URLs in the Firefox helper.
I keep trying them and being surprised when they don't work. It causes
me to disable the helper and run using just meek-client instead. The
restriction against http URLs was intended as a failsafe, to prevent
people from accidentally configuring an http URL when they meant to do
https. But I think its value as a diagnostic outweighs that
consideration.