Reuse the same RoundTripper for all requests.
Previously we were creating a new http.Transport for each request, and
selectively overriding the Proxy member. Since that code was first
added, http.DefaultTransport has gotten other non-zero settings than
Proxy: Dial and TLSHandshakeTimeout.

https://code.google.com/p/go/source/diff?spec=svn733fefb1deae5490f2d6fcf498667a9077f92a42&name=733fefb1deae&r=733fefb1deae5490f2d6fcf498667a9077f92a42&format=side&path=/src/pkg/net/http/transport.go
https://code.google.com/p/go/source/diff?spec=svnc8edfe4ddd5a28d929503ac34d3918759518970c&name=c8edfe4ddd5a&r=c8edfe4ddd5a28d929503ac34d3918759518970c&format=side&path=/src/pkg/net/http/transport.go

We want to use those settings, but we also want to disable the default
ProxyFromEnvironment. Creating a new Transport every time was messing
with persistent connections; I watched as meek-client made a new TCP
connection for every request. With this change, it reuses the same
connection. Something must have changed in the http library (perhaps one
of the revisions cited above), because I remember it using a persistent
connection in the past.

This only makes a difference when not using --helper. The browser helper
handles all its own persistence stuff.