Don't use the SOCKS requested address as a fallback URL.
The code was set up to use the SOCKS address in case neither a url=
SOCKS arg nor the --url option was present. So that this:
	Bridge meek 192.0.2.1:80
would in effect be the same as
	Bridge meek 192.0.2.1:80 url=http://192.0.2.1:80/
This feature was probably not ever used except as the result of a
misconfiguration. It's better to remove it than be surprised by it
later. It also lets us state unequivocally that the SOCKS address is
always ignored.