Reduce meek-server minSessionIdLength from 32 to 8.
Relevant to #12778, we can reduce the size of requests by reducing the
length of the Session-Id. Note that minSessionIdLength in meek-server is
the encoded length. In base64, 8 characters represents 6 bytes or 48
bits. There would have to be about 2^24 simultaneous sessions to have a
probability of collision more than half, so it still seems safe.