Read data before bailing out in onDataAvailable.
https://developer.mozilla.org/en-US/docs/Mozilla/Tech/XPCOM/Reference/Interface/NsIStreamListener#onDataAvailable%28%29
says "Your implementation of this method must read exactly aCount bytes
of data before returning."

It's already downloaded, so it doesn't cost much to read it out of the
buffer. Without this, I got "http channel Listener OnDataAvailable
contract violation" in the console. (I had to purposely induce an
overlong download to get it to happen.)