Use persistent connections in the WSGI reflector.
This improves performance quite a lot--previously we were doing a
complete TCP and TLS handshake to meek-server for every single request,
which, apart from increasing latency, also caused a lot of CPU usage on
meek-server. it was up above 80% when I checked it.

Now we reuse connections until they error out, making a new connection
if all others are currently busy.