Set scaling instances in app.yaml.
These used to be set using the web interface. Now you have to set them
in app.yaml. For example,
https://cloud.google.com/appengine/docs/adminconsole/performancesettings#Setting_the_Number_of_Idle_Instances
now redirects to
https://cloud.google.com/appengine/docs/developers-console/#module-settings
which says:
	Note: Many performance settings for modules are included in the
	module's configuration file (Java | Python | Go | PHP). These
	settings are made at deployment time and cannot be changed from
	the Developers Console.
So we've been running without the settings I set manually before for who
knows how long. Previously I had set:
  max_idle_instances: 2
  min_pending_latency: 1000ms
Now I set it to:
  max_idle_instances: 2
  min_pending_latency: 1000ms

The new URL for the documentation of these parameters is here:
https://cloud.google.com/appengine/docs/go/modules/#Go_Configuration