Ignore SIGPIPE in meek-client-torbrowser.
See https://bugs.torproject.org/20030 for discussion. Since Go 1.6,
writes to fd 1 or 2 when they are closed causes the program to terminate
with a SIGPIPE. Because the default log location is stderr (fd 2), we
would get a SIGPIPE when logging once the parent process had died and
closed its stderr. This prevented meek-client-torbrowser from cleaning
up its subprocesses.