Integrate uTLS as an option for TLS camouflage.
This adapts a technique that Yawning used in obfs4proxy for meek_lite:
https://gitlab.com/yawning/obfs4/commit/4d453dab2120082b00bf6e63ab4aaeeda6b8d8a3
https://lists.torproject.org/pipermail/tor-dev/2019-January/013633.html

It's activated by the new utls= SOCKS arg or --utls command line option.
The argument is the name of a uTLS ClientHelloID; e.g.,
HelloChrome_Auto. We omit HelloCustom (not useful externally),
HelloGolang (just don't use utls), and HelloRandomized (may negotiate
HTTP/1.1 or HTTP/2 at different times, which is incompatible with the
way the integration works).

When using HTTP/1, we copy default timeouts etc. from the Go
net/http.DefaultTransport. Unfortunately I don't know a way to do the
same for HTTP/2. Configuring an http.Transport and then calling
http2.ConfigureTransport on it doesn't work; it leads to the same
problem of an HTTP/1 client speaking to an HTTP/2 server.

We recognize utls=none and utls=HelloGolang as an alias for omitting
utls=. This is for compatibility with obfs4proxy meek_lite.
https://bugs.torproject.org/29077#comment:13