Have terminateprocess-buffer wait on either of stdin closing or child exiting.
On Windows, I observed once that if meek-client.exe can't be found,
tor.exe and its child process terminateprocess-buffer.exe continued
running. I suspect terminateprocess-buffer was hung up in io.Copy--the
child process stdin pipe was closed, but as it has nothing to copy, it
couldn't detect the fact. Now have it do the stdin copy in a separate
goroutine, and also close the child stdin pipe explicitly.